RM-83432 remove a duplicate file
